arp_pipeline/pyproject.toml
[tool.poetry]
name = "arp_pipeline"
version = "0.1.0"
description = ""
authors = ["Matt Bowen <matt@mattbowen.net>"]

[tool.poetry.dependencies]
python = "^3.8"
luigi = "^3.0.3"
SQLAlchemy = "^1.4.25"
psycopg2 = "^2.9.1"
requests = "^2.26.0"
plumbum = "^1.7.0"
python-dotenv = "^0.19.0"
pandas = "^1.3.3"
openpyxl = "^3.0.9"
GeoAlchemy2 = "^0.9.4"
pyarrow = "^5.0.0"
platformdirs = "^2.4.0"
CensusData = "^1.15"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.9b0"
ipython = "^7.27.0"
pylint = "^2.11.1"
isort = "^5.9.3"
types-requests = "^2.25.8"
autoflake = "^1.4"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tools.black]
line-length = 119
exclude = '''
\(
    \.git
    | \.mypy_cache
    | \.tox
    | venv
    | \.venv
    | _build
    | build
    | dis
    | data
\)
'''

[tools.isort]
profile = "black"

[mypy]
ignore_missing_imports = true
